+++ src/lib/packages/package.ts
@@ -11,7 +11,9 @@ import ResourceEntry from "./resource-entry";
  * Model for a Sims 4 package file (also called a "Database Packed File", or
  * DBPF for short).
  */
-export default class Package extends MappedModel<ResourceKey, Resource, ResourceEntry> {
+export default class Package<ResourceType extends Resource = Resource>
+  extends MappedModel<ResourceKey, ResourceType, ResourceEntry<ResourceType>> {
+
   //#region Initialization
 
   /**
@@ -19,8 +21,10 @@ export default class Package extends MappedModel<ResourceKey, Resource, Resource
    * 
    * @param entries Entries to initialize Package with
    */
-  constructor(entries?: ResourceKeyPair[]) {
-    super(entries); // never cache package buffer, it's pointless
+  constructor(entries?: ResourceKeyPair<Resource>[]) {
+    // intentionally using Resource in arguments so that ResourceType isn't
+    // automatically inferred from the first element
+    super(entries as ResourceKeyPair<ResourceType>[]);
   }
 
   /**
@@ -29,8 +33,11 @@ export default class Package extends MappedModel<ResourceKey, Resource, Resource
    * @param buffer Buffer to read as a package
    * @param options Options for reading and cacheing the resources
    */
-  static from(buffer: Buffer, options?: PackageFileReadingOptions): Package {
-    return new Package(readDbpf(buffer, options));
+  static from<T extends Resource = Resource>(
+    buffer: Buffer,
+    options?: PackageFileReadingOptions
+  ): Package<T> {
+    return new Package<T>(readDbpf(buffer, options) as ResourceKeyPair<T>[]);
   }
 
   /**
@@ -39,8 +46,11 @@ export default class Package extends MappedModel<ResourceKey, Resource, Resource
     * @param buffer Buffer to read as a package
     * @param options Options for reading and cacheing the resources
     */
-  static async fromAsync(buffer: Buffer, options?: PackageFileReadingOptions): Promise<Package> {
-    return promisify(() => Package.from(buffer, options));
+  static async fromAsync<T extends Resource = Resource>(
+    buffer: Buffer,
+    options?: PackageFileReadingOptions
+  ): Promise<Package<T>> {
+    return promisify(() => Package.from<T>(buffer, options));
   }
 
   /**
@@ -50,27 +60,33 @@ export default class Package extends MappedModel<ResourceKey, Resource, Resource
    * @param buffer Buffer to extract resources from
    * @param options Options for reading and cacheing the resources
    */
-  static extractResources(buffer: Buffer, options?: PackageFileReadingOptions): ResourceKeyPair[] {
-    return readDbpf(buffer, options);
+  static extractResources<T extends Resource = Resource>(
+    buffer: Buffer,
+    options?: PackageFileReadingOptions
+  ): ResourceKeyPair<T>[] {
+    return readDbpf(buffer, options) as ResourceKeyPair<T>[];
   }
 
   /**
-   * Asynchronously reads the given buffer as a Package, but just returns its entries rather
-   * than a full Package object.
+   * Asynchronously reads the given buffer as a Package, but just returns its
+   * entries rather than a full Package object.
    * 
    * @param buffer Buffer to extract resources from
    * @param options Options for reading and cacheing the resources
    */
-  static async extractResourcesAsync(buffer: Buffer, options?: PackageFileReadingOptions): Promise<ResourceKeyPair[]> {
-    return promisify(() => Package.extractResources(buffer, options));
+  static async extractResourcesAsync<T extends Resource = Resource>(
+    buffer: Buffer,
+    options?: PackageFileReadingOptions
+  ): Promise<ResourceKeyPair<T>[]> {
+    return promisify(() => Package.extractResources<T>(buffer, options));
   }
 
   //#endregion Initialization
 
   //#region Public Methods
 
-  clone(): Package {
-    return new Package(this.entries.map(entry => entry.clone()));
+  clone(): Package<ResourceType> {
+    return new Package<ResourceType>(this.entries.map(entry => entry.clone()));
   }
 
   equals(other: Package): boolean {
@@ -85,7 +101,10 @@ export default class Package extends MappedModel<ResourceKey, Resource, Resource
     return `${key.type}_${key.group}_${key.instance}`;
   }
 
-  protected _makeEntry(key: ResourceKey, value: Resource): ResourceEntry {
+  protected _makeEntry(
+    key: ResourceKey,
+    value: ResourceType
+  ): ResourceEntry<ResourceType> {
     return new ResourceEntry(key, value, this);
   }